Reload session on the finally callback
