+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/mobile/MobileActiveController.kt
@@ -49,4 +49,8 @@ class MobileActiveController(
     override fun onDeleteSessionClicked(sessionUUID: String) {
         // do nothing
     }
+
+    override fun onExpandSessionCard(session: Session) {
+        // do nothing
+    }
 }
