+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/mobile/MobileDormantController.kt
@@ -6,6 +6,7 @@ import androidx.lifecycle.LiveData
 import io.lunarlogic.aircasting.database.DatabaseProvider
 import io.lunarlogic.aircasting.database.data_classes.SessionWithStreamsAndMeasurementsDBObject
 import io.lunarlogic.aircasting.events.DeleteSessionEvent
+import io.lunarlogic.aircasting.exceptions.ErrorHandler
 import io.lunarlogic.aircasting.lib.Settings
 import io.lunarlogic.aircasting.models.ActiveSessionsObserver
 import io.lunarlogic.aircasting.models.DormantSessionsObserver
@@ -13,6 +14,7 @@ import io.lunarlogic.aircasting.screens.dashboard.SessionsController
 import io.lunarlogic.aircasting.models.SessionsViewModel
 import io.lunarlogic.aircasting.screens.dashboard.SessionsViewMvc
 import io.lunarlogic.aircasting.models.Session
+import io.lunarlogic.aircasting.networking.services.DownloadMeasurementsService
 import org.greenrobot.eventbus.EventBus
 
 class MobileDormantController(
@@ -45,22 +47,5 @@ class MobileDormantController(
 
     override fun onStopSessionClicked(sessionUUID: String) {
         // do nothing
-    }
-
-    override fun onExpandSessionCard(session: Session) {
-        super.onExpandSessionCard(session)
-
-        if (session.isIncomplete()) {
-            DatabaseProvider.runQuery { scope ->
-                val dbSessionWithMeasurements = mSessionsViewModel.reloadSessionWithMeasurements(session.uuid)
-                dbSessionWithMeasurements?.let {
-                    val reloadedSession = Session(dbSessionWithMeasurements)
-
-                    DatabaseProvider.backToUIThread(scope) {
-                        mViewMvc.reloadSession(reloadedSession)
-                    }
-                }
-            }
-        }
     }
 }
